Address comments. Move auth and healthcheck inside nginx package
