+++ ingress/controllers/nginx/.gitignore
@@ -1,0 +1,1 @@
+nginx-ingress-controller
